Implement tvwrite.cpp
This code has been tested with the TVDEMO application in both 16-bit and 32-bit. There were no issues so far. Unlike tvexposed, I haven't made a runtime comparison against the assembly version.

I first tested this code with the TVDIR application; I experienced crashes when closing the application and had lots of trouble figuring out the reason. It turned out to be a very long filename in one of the Windows XP system folders that caused stack corruption. This resulted in a crash in TOutlineViewer::disposeNode (toutline.cpp), where the nodes of the tree used to represent the filesystem are deleted.

TVDEMO was the second application I tested this against. At first it crashed upon creation of the color editor window, which turned out to be caused by a 'return True' I had left many time ago in TView::exposed(); secondly, it crashed after closing the Help dialog, but that's caused by the changes I made to the part of the code that deals with opening files.