Fix crash when editing text
Amazing. The assembly function lineEnd, which searches for a carriage return in the current line, uses the REPNE instruction which relies on the (E)CX register to stop iterating. When they wrote this function in 32-bit assembly, though, they forgot about the upper half of ECX. Therefore, if the upper half of ECX was different from zero, REPNE would keep iterating even if CX was zero, which is how it worked in 16-bit.

For some reason (either due to a huge amount of luck, or because they knew too well what the program was doing) this still worked if the assembly version of TVWRITE was in use; the assumption that ECX<31..16> equals zero is no longer true when using the new CPP version. This made me initially believe this was a bug in the code I wrote, but it is not.