Prevent static initialization order fiasco
The original design of Turbo Vision initializes all system interfaces statically, which makes any extensions to it vulnerable to a static initialization order fiasco.

This can be avoided by constructing the interfaces on their first use when constructing TApplication. The downside of this solution is that system interfaces cannot be used before the TApplication constructor is ran for the first time (which isn't necessarily a flaw, but I can't be sure that no applications rely on that behaviour).

In any case, this can always be refined or made optional in the future. With the current changes, TVDEMO still works well under Windows and DOS.