Fix regression in THelpViewer
cd793fbdbb44d3e28667ffdb1182ef7446713448 replaced fixed-length arrays with dynamic ones in THelpViewer::draw(). What I did not see back then was that part of the code relied on sizeof to measure the length of the array, which works fine in the original code but returns the size of a pointer after the changes.

This caused THelpViewer to trim the text it displayed and append garbage next to it, in addition to segfaulting when deleting the affected array on certain environments.