Add support for GPM mouse on the linux console
The LinuxConsoleStrategy class might seem a bit overkill for managing two AsyncInput instances, and indeed it could be implemented as a composition in GpmInput instead, but we need this for future extensions.

shared_ptr attributes have been replaced with unique_ptr because shared_ptr requires the use of make_shared and passing shared_ptrs in constructors  to work properly. While we don't need a more complex pointer management, we can just use unique_ptr to express pointer ownership.