AsyncInputStrategy: use lazy thread creation and automatize detaching
In order to prevent data races between the main thread and the input threads before resumeListening() is reached for first time, listeners register themselves upon construction and get started automatically the first time resumeListening() is called.

With these changes, overriding getEvent needs to be done before the thread is started. We accomplish this by storing an eventGetter functor that can be then overriden.

Also, in order to make their usage simpler and more RAII-ish, threads get detached in the destructor.