TFileDialog: implement getFileName()
Rather than writing to a buffer, this function returns a string that has the same lifetime as the TFileDialog object.

To manage this, a 'fullPath' attribute has been added to TFileDialog where this string can be stored.

tvdemo has been updated to use this new implementation.

A fallback version limited to MAXPATH characters has been added in order not to break the public interface of TFileDialog.