TNSCollection: use realloc to increase performance for large collections
An example of this is TLineCollection when reading a long text file or a binary. Each time the collection limit is exceeded, the 'items' list is copied into a newly allocated area, then the new elements are inserted. However, this is order O(n*n).

This can be handled more efficiently with realloc/free.