THistoryWindow: close history window when clicking outside it
This is a usability necessity.

It turns out that it was just necessary to override TWindow::handleEvent from THistoryWindow. This was not obvious to me, as there's also a THistoryViewer class, and all that can be seen from TFileDialog's code is that it creates an instance of THistory. It took me some backtraces to get the big picture.