TListViewer: fix highlighting delay when selecting entries by mouse drag
In the original do-while loop, the highlighted element was drawn before processing the last mouse event. Thus, there was a one-event delay.

This can be fixed by redrawing the view at the end. To do so, we must not process the first event separately outside the loop (or duplicate more code).

I can't think of a situation where the first mouse event is outside the view's bounds, so I don't see the advantage of doing it that way (the else clause here is different from the one inside the loop, but the code there is not likely to break anything).

Thus, there's a bunch of duplicated code we can avoid.