NcursesInput: do not discard repeated events on mouse drag
Actually, this is what evMouseAuto consists in. Unfortunately, not all terminal emulators have this behaviour, and it is not clear whether it should be considered a bug or a feature.

For now, it works in Konsole, so I can have TFileList extend its selection when dragging the mouse out of its bounds. Otherwise, we could try to change the way evMouseAuto is detected (in TEventQueue::getMouseEvent).

These kind of events are already not filtered out (I believe) in GpmInput, but it's Gpm which doesn't usually report repeated mouse events.