TMenuView: hide submenus by clicking on their names
Much better now! :)

It took me some time to figure out how to do this.

In first place, mouseInOwner() had to be modified to work for parent menus which are not a menu bar. This was explicitly disabled for some reason, but removing the specific check seems to work.

MouseDown events can then cause a submenu to close when the click fell on the parent menu. The parent menu must not open this submenu automatically again, which was the default behaviour of the menu bar because of autoSelect being True. It should be False when the menu bar has a highlighted entry and it's also the last one to be executed. When no entries are highlighted, setting autoSelect to False would render the menu bar unable to open any submenus via mouse click.

MouseUp events would open again a closed menu by default. So a check had to be added to prevent this from happening.

Finally, dragging the mouse after closing a submenu should not prevent a MouseUp event from reopening that submenu again if it became unselected at least once.

This solves most of #6.