TMenuView: do not select menu bar's default entry on MouseUp over empty area
Because menu bars have autoSelect = True, the entry would open up unexpectedly after clicking on an empty area. The fact that the default entry is not very consistently chosen made this look quite broken.

However, this is pretty useful for submenus, which don't have autoSelect, so the functionality must be kept active for that case.

Closes #6.