TScrollBar: handle evMouseWheel by default
I'm not very familiar with Turbo Vision events yet, but I can see that events which are neither 'positional' or 'focused' (see the change in views.h) reach all the views of a TGroup (see TGroup::handleEvent()).

Ideally, the window holding scrollbars would catch the event and apply a specific strategy, e.g. make a horizontal scrollbar handle evMouseWheel if there's no vertical bar, or decide which scrollbar should handle it if there are more than one.

But rather than touching code everywhere, let's just allow scrollbars to catch mouse wheel events and handle them if required. This simply works out-of-the-box for any window with a single vertical scrollbar, such as the file viewer in tvdemo or TEditorWindow. Horizontal bars will only handle horizontal scrolls and vertical bars will only handle vertical scrolls. This can be extended in the future so that scrollbars may override this restriction.

Excluding evMouseWheel from positionalEvents (views.h) makes more sense than redefining evMouse not to include it. evMouse is often used in do-while(mouseEvent()) loops, so it is desirable to have mouse wheel events catched by these, I guess.

This is enough to close #5 for now.