Add interface support for screen resizing
This commit adds two definitions: the message 'cmScreenChanged', to warn TProgram that the screen size changed, and the video mode 'smChanged', which allows TScreen to allocate a new buffer which matches the current screen dimensions when calling TScreen::setVideoMode, without having to alter code elsewhere.

The key function is TProgram::setScreenMode, which does all the necessary stuff, but is a member function and can't thus be called from getKeyEvent without a TProgram object. Therefore, getKeyEvent, after capturing a screen resize event, will be able to return a cmScreenChanged message which will be first read by TProgram.