Windows: handle screen resizing events
On Windows, only buffer size changes are reported to the program. That is not an issue. The buffer size may be changed through the console's 'Properties' dialog. wineconsole also supports enlarging the buffer by resizing the console window.

I've refactored GetConsoleScreenBufferInfo into a new 'reloadScreenBufferInfo' function, to make the process more abstract. But I haven't touched the invocation in THardwareInfo's constructor, because it helps understand the startup process.