Fix crash when returning from shell in wineconsole
wineconsole sometimes resizes the screen buffer in a way that TScreen and THardwareInfo's structures are up to date but TApplication stays unaware, and views attempt to draw beyond TDrawBuffer's limits, because TDrawBuffer retrieves screen dimensions directly from TScreen.

Well, it used to retrieve them from TDisplay, but I've changed my mind. This seems more consistent.

The only changes that actually helps prevent the crash are the ones in TApplication::resume(). Calling reloadScreenBufferInfo() in THardwareInfo::setUpConsoleBuffer() is unnecessary but I added it just in case.

Closes #5.