Add new single-thread InputStrategy interface
Running input retrievers in separate threads sounds cool, but is vulnerable against data races in the libraries being used (Ncurses is not thread-safe) and prevents SIGWINCH from reaching to them (in case they can handle it).

The obvious single-thread solution is to take note from the file descriptors being read by input listeners, then detect when data is avaiable in them with a system call.

This can still be done in a distributed fashion. Instances of this class should call 'addListener' to have their 'eventGetter' called when a certain file descriptor is ready to read.