Add support for capturing SIGWINCH signals
The new FdInputStrategy allows input retrievers to work in a single-thread manner. Thanks to this, Ncurses can now capture SIGWINCH, but that won't place any data into stdin, so FdInputStrategy will be unable to realize NcursesInput should be invoked.

The solution is to set a custom handler function for SIGWINCH which will not only call the previous handler (the one set by Ncurses in our case), but also write to a file descriptor. The new SigwinchAware class should be inherited by InputStrategy implementations able to handle this signal, and then call 'addListener' passing the readable file descriptor provided by this class. As a result, FdInputStrategy will know they have to be invoked after a SIGWINCH signal is catched.

SigwinchAware cannot work individually. In the case of Ncurses, doing so would prevent it from realizing the screen dimensions changed.