Warn DisplayStrategy after a screen resize
DisplayStrategies may need to know when the screen resolution changed. In the case of BufferedDisplay, this is necessary to resize the internal buffer.

In order to do this, we add a new function THardwareInfo::resizeScreenBuffer that gets called by TScreen after a change in screen resolution was noticed. This function calls PlatformStrategy::onScreenResize, which reaches the current DisplayStrategy.