Fix received vs processed SIGWINCH conflict
The original design assumed Ncurses would always return KEY_RESIZE as many times as SIGWINCH was catched; by not doing so, the pipe could be left with remaining data with Ncurses not returning KEY_RESIZE anymore, resulting in 100% CPU usage.

A better approach is to write at most one element into the pipe and always return a cmScreenChanged command if Ncurses invokes winchEvent. Then, emptying the pipe is done whenever KEY_RESIZE is not returned, thus being agnostic about the count of catched SIGWINCH and emitted KEY_RESIZE, and sticking to the only assumption that KEY_RESIZE will be the only event returned by getch() whenever Ncurses is aware of a change in screen resolution.

In case this failed, it would not run into 100% CPU usage. I haven't experienced any problems with this so far.