Fix cmScreenChanged not being processed with active windows
TProgram::handleEvent is not guaranteed to be always called after retrieving an event. This is the case when handleEvent is called from TGroup::execute.

The solution is not very difficult: TProgram::getEvent IS always called in a sane application design. That's where the handling of cmScreenChanged has to be done.