Linux: release/grab console control when suspending/resuming
THardwareInfo's functions 'setUpConsole' and 'restoreConsole' are in the right place to handle these situations. The changes also prevent the existence of more than one PlatformStrategy instance at the same time, which would cause terrible mess with GpmInput as FdInputStrategy doesn't check for duplicate file descriptors and Gpm_GetEvent is a blocking call.