tdircoll.cpp: validFileName: fix files being undesirably created
And I thought using an ofstream would be a good idea...

After looking at it for a second time, it looks like it is enough for validFileName to check the parent directory exitst. There are no illegal characters in linux filesystem names, and the underlying system calls should be responsible of checking everything is fine.