TMenuView: allow opening last selected submenu after cmMenu event
lastTargetItem prevents a submenu from opening. It's the desired behaviour when a submenu is closed by clicking on its name; instead, when a cmMenu event is received (e.g. by pressing F10 or clicking on the 'Menu' status bar entry), it should be possible to open it again anyway.

This is actually a regression of 287c4f392bcaea834b46ef6f445bb13de5ff68a1. Related to #6 as well.