Fix views losing their size when extremely resized
Views are not allowed to be resized out of the limits provided by sizeLimits, which is usually a range between zero and the size of the owner. When fitting the new boundary within these values, views keep no track of its original size.

This is a trouble when resizing the terminal. Even if a window has no wfGrow flag, it won't be allowed to be greater than the desktop, and will be shrinked if necessary, including all its subviews.

To ensure everything goes back to its original size when the terminal is again big enough, a balance of the remainders has to be stored for each view. If the suggested bounds are outside the allowed range, then that difference is added to the balance. If they are inside the range, then as much as possible from the balance is added back to the boundaries.

This commit also applies a minor change to the grow() macro. In the gfGrowRel branch, now it just checks that s and d are different and does nothing otherwise.