Windows: exit if not running under a Win32 console
This prevents the program from consuming 100% CPU (because console I/O operations fail) and becoming unexitable (as it can't be killed by neither Alt+X or Ctrl+C). This would be the case, for example, with the MINGW (Git BASH) terminal, or when redirecting stdin. Redirecting stdout is not a problem as WriteConsoleOutput doesn't rely on it.