Windows: fix reading of Ctrl+C and Shift+Arrow events
Ctrl+C was clearly being captured by the CtrlBreakHandler, but Shift+Arrow events were being processed by Conhost for no known reason. It turns out this all comes down to the ENABLE_PROCESSED_INPUT flag.

Fixing this has been a hassle because of bad programming practices during testing. I wanted to print the actual values of the console mode for stdin and stdout, and when retrieving these I reused the consoleMode variable, without taking into account that it is later used in TMouse::cursorOn. So, without me knowing, the application was running with the input console mode set to the same value as the output one.

The reason why I'm explaining this? I've been wanting to fix this for so much time it hurts me deeply.