FdInputStrategy: fix 100% CPU usage when stdin is disconnected
This can happen, for instance, if the application is run with root permissions but the terminal is not. When closing the terminal will be unable to kill the application and stdin will become broken. This makes poll also wait in such situations, and even if no file descriptors are being listened at all.