Implement buffering in TTextDevice
Drawing characters in an unbuffered TTextDevice is a very expensive operation, as do_sputn is called for every written character, which in turn calls drawView().

Buffering could be implemented by inheriting from the deprecated strstreambuf, but we would have to define xsputn (the equivalent of do_sputn in the standard) and deal with unavoidable flushes for every '<<' operation in a ostream (thus having less performance than we'd like).

So I have found it preferable to do a very basic implementation of buffering based on the streambuf interface. TTextDevice now supports the sync operation, so std::flush or std::endl will be required to draw the buffered data.

This improves performance greatly in the event viewer.