TTerminal::prevLines: fix spontaneous printing of '\n'
When exiting the main 'if' block, pos is -1 at least. The only way pos + 2 can be greater than bufSize is because of exceedance and not arithmetical overflow. This is a circular buffer, so when bufSize is exceeded, it has to be subtracted.

In order to improve clarity (or not) and prevent any possible out-of-bound memory access, the last character read from buffer has been backed up in a local variable.

The issue appeared in cases where 0 was returned because pos + 2 was being read as 65535 + 2 instead of 1, and position 0 in the buffer happened to contain a newline character.