Do not use string.h for legacy string functions
Many of these functions were defined in stdlib.h and not string.h. Anyway, the main purpose of this is to get rid of string.h, as it leads to compilation issues with compilers other than GCC.

This also increases the buffer size in TFileInfoPane::draw(): since longs are 64-bit on linux, 10 bytes are no longer enough to hold all digits of a long.