BufferedDisplay: replace Set with better strategy
The ordered set performs badly. It is better to store all character information in a single matrix, which takes advantage of data locality and does not require memory allocations during buffering.

Removing the set, though, prevents us from having a bounded list of modified cells. In order to reduce the number of iterations in flushScreen(), we keep damage information for every screen row.