Change lowlevelWriteChar interface
The call to CpTranslator::toUtf8 has been moved to BufferedDisplay to avoid duplicated code. Additionally, we are now officially ignoring the upper byte of character attributes, so the attr argument can be changed to uchar.

This commit also undoes a change in tables.cpp that replaced null characters with spaces. It resulted in spaces being translated as null on input. Move that check back to BufferedDisplay.