Definitely ensure inheritors of trivially_convertible are trivial
The failing assertion was muted in c0807690786306f02900d14089520b97c880c3d0 but that resulted in Valgrind reporting unexplainable usages of uninitialized data. I don't know whether this is a bug in Clang or a misconception of mine, but manually defining a trivial default constructor in these classes fixes the issue.