Redesign TDrawSurface interface
There is no point in TDrawSurface (formerly TSurface) being a TView. TSurfaceView already takes care of drawing it. This way the only responsability of TDrawSurface is storing a buffer and resizing and filling it when requested to do so. It is more similar to a TDrawBuffer than a TView.

This commit also implements support for all kinds of deltas in TSurfaceView, so there is no need for the delta setter anymore. The surface setter has also been removed because there is no need for it.