THelpViewer::draw: make more room for lines
The use of a fixed-length buffer is unavoidable because of getLine's interface. However, we previously made this buffer's length proportional to the screen's width, which does not make much sense. A size of 256 is greater than TDrawBuffer's length in most cases, and it won't be dangerous as long as the buffer is kept null-terminated, which we do.