Exterminate fixed-length char arrays in THelpTopic
Holy shit. Not only were all these char copies unnecessary in many cases (getCrossRef and getNumCrossRefs do not read the text), but they were also useless: the desired text can be accessed with zero overhead with a mere TStringView. Yes, I know that string views may become dangling pointers, but in our case the lifetime of the THelpTopic object exceeds the scope of usage of the result of THelpTopic::getLine (in THelpViewer::draw). This is yet another example of the harm null-terminated strings have caused to programs and programmers.

Fixed-length arrays were also (again) introducing stupid limitations, such as the hardcoded maximum length of 256 in helpbase.cpp:scan. Heck, there was even an unnecessarily duplicated invocation to strlen in THelpTopic::wrapText. Shame.