tvhc: fix semicolon insertion in output header
The original implementation seeks back 3 bytes to skip ",\r\n". Since the line endings differ between platforms, the simplest solution is not to depend on this implementation detail. By adding newlines at the beginning of each symbol definition (instead of the end), the only character that has to be overwritten is the comma.