Fix compilation with older CMake/GCC and link in tinfow
The 'if (NCURSESW)' has been removed because ncurses is already compulsory. The REQUIRED option of find_library is not available in older CMake versions, so the only way to trigger an error is to dereference NCURSESW unconditionally.

tinfow has to be linked in systems where both tinfo and tinfow are available. See #11.