Improve gitignore (2)
It is safe to ignore all directories by default because if a new directory with source code is added, it is unlikely that 'git add'ing it will become a bothersome task. On the contrary, it is very likely that all sorts of temporary directories will be created and I cannot predict all of their names.