Definitely fix linkage of TStreamableClass initializers
It turns out that the 'volatile' workaround of 281db91e38b182c9f8667e534cefb0078ffb9cb9 does not work for all compilers. By having a global volatile variable, and writing to it on every usage of the __link macro, we prevent the compiler from optimizing out the references with negligible overhead, even with LTO enabled.

Since the variable is meant to be write-only, we do not have to care about thread safety nor static initialization order.

Note that we do not need a struct like fLink just to write to a global variable. But since it has a reference to itself, unused static fLink variables do not produce a warning on Borland C++ and possibly other compilers. Additionally, in case anyone was manipulating fLink structs manually (which I doubt), backward compatibility is preserved.