Moderate usage of constexpr map
After some testing, I found that the performace of constexpr_map is not better than unordered_map for not-so-small map sizes, which is moreless the case of the tables of NcursesInput. So these have been replaced with const unordered_maps. The tables in LinuxConsoleStrategy are very small, so they have been left constexpr.