Signal handler: `exit` instead of `signal`
This implies dealing with the situation where THardwareInfo's functions are invoked during the destruction of TEventQueue and no PlatformStrategy is loaded (I wonder why this isn't an issue when the program ends normally).

We address this issue by having a default no-op implementation of PlatformStrategy, instead of adding 'ifs' everywhere.