Undo screenWrite performance improvements
Further testing shows that the benefits can only be experienced with GCC, and that performance regresses severely with Clang.
There still are a few improvements that can be done, such as caching 'screenChanged' in a local variable. This alone already provides small performance improvements that render the previous invasive changes useless.

This reverts commit b40cd033ff62b61a16c06abf3aca5f539c0d60c6.