TDrawBuffer: always have length >= 80
With extreme screen sizes, it's easy to run across a view that's wider than the screen itself, e.g. menu boxes. These view won't fit into TDrawBuffers and will be drawn incompletely. The definitive way to solve this is by sizing TDrawBuffers according to the view using them, but the current interface doesn't allow this. So, as a mitigation, make all TDrawBuffers at least 80 columns wide.