TInputLine: do not scroll text display on focus/unfocus
When a TInputLine is clicked on, it becomes focused before processing the mouse  event. This resulted in the click not matching the location the user was seeing.

Similarly, when unfocusing a TInputLine, the display was scrolled to the beginning, making it impossible to see the text that was previously visible.

This commit prevents scrolling to happen in both focus and unfocus events. This will require changes in inheriting classes that rely on this.

In the case of TFileInputLine, for which it is convenient to scroll to the beginning when a new file is selected in a TFileList, this is done from the same place where text is updated, TFileInputLine::handleEvent.

This fixes the issue that 57a5957d9069c358da789b18f738cd436814a8d5 failed to solve.