TEvent: make conversion to TStringView explicit
Implicit conversions are cool, but explicitness is sometimes good. In this case, it makes it easier to construct a std::string_view or std::string from a TEvent, because TStringView can implicitly cast to them. But with an implicit cast, `operator TStringView` would only be invoked when the result is being assigned to a TStringView variable. Additionally, having several implicit casts (which is not the case here) can easily cause conflicts. So a explicit function makes things clearer and is more future-proof.