Add support for displaying zero-width characters
This turned out to be a lot easier than I anticipated. Zero-width characters simply get combined into the previous cell in TText::eat, or are not shown at all. The zero width joiner is always discarded so that it won't merge characters together, changing the width of a whole string.

In order to be able to combine several characters in the same screen cell, the size of TCellChar has been raised to 8 bytes. This allows a minimum of one combining character, which should be fine for most non-degenerate use cases. But since we use UTF-8, several codepoints may fit in there. It also preserves the assumption that TCellChar can be casted into a primitive type, so there's not much code that has to be changed elsewhere.

The only breaking change is TText::eat, where the 'width' and 'bytes' parameters are now indexes so that we can look back at the previous cell when we find a zero-width character. But this actually makes things simpler for the invokers of TText::eat. So it's a win-win change.

See #26.