Implement TVMemMgr::reallocateDiscardable
According to https://stackoverflow.com/a/39562813, realloc is better than free + malloc when shrinking, but worse when growing. So let's create a new function in TVMemMgr that encapsulates this logic, and use it to resize the buffer in TGroup.

This may provide a small performance improvement and memory saving.

There is only one thing I'm not sure about: the `extra*sizeof(unsigned)` in `void *TBufListEntry::operator new(size_t, size_t)`. This increases memory consumption unnecessarily; it may be that it was placed there during debugging and the authors forgot to remove it, or that they were suffering from out-of-bounds memory access bugs and allocated extra memory to work around them. In any case, let's revert this overallocation and see what happens.