Redesign non-blocking input feature
After having confirmed that this feature was feasible, it's time to improve it. The initial interface was terrible, because it forced the caller to manage the event queue. Since the only reason why we are doing this is to read text from input, `getImmediateText` has been replaced with `getTextEvent`, which does most of the work. The only thing the caller is now responible of is memory management. I considered returning a dynamically allocated array of char, but this is dangerous, as it can lead to unbounded memory usage. To avoid this we could limit the maximum readable size, but then why not just let the caller manage this themselves. Additionally, reading in moderately sized chunks is also the most efficient way to do this kind of operation. So I decided there was no advantage in forcing dynamic allocation. I acknowledge reading by chunks implies relying on a global state, but the only solution to this would be to use callbacks, and Turbo Vision is not ready for this.