Do not use CMAKE_PROJECT_NAME as target name
This variable is influenced by superior scopes, so it won't work if including Turbo Vision with add_subdirectory. It turns out that the variable PROJECT_NAME does what we need, but I can see many popular libraries not encapsulating the target name in a variable, so I guess we don't need to do it either.