NcursesInput: add support for parsing mouse events manually
This was a lot easier to integrate than I anticipated. However, the logic is not simple because the parse functions may be invoked from two different places (parseEscapeSeq and parseCursesMouse), and we must perform trial-and-error to figure out the actual encoding. Anyway, this may solve many issues with mouse input.

This approach may be controversial, though, as it overlooks the terminal capabilities database.

Fixes #29.
Fixes #31.