Specialize ScreenCursor statically
When I first wrote this class I thought that inheritance was overkill for the default styles, so I added a flag attribute to differenciate them, but still kept the draw() method virtual so that it could be inherited. This doesn't make much sense, though, as any other specialization would have to deal with the ScreenCursor constructor taking an unused parameter.

Simply specialize in separate classes. This also reduces complexity in the classes making use of ScreenCursor, i.e. GpmInput.