TApplication: provide default implementation of dosShell(), tile() and cascade()
These commonly used methods are defined in the example applications and most people will copy them from there, leading to duplication. This is especially bad in the case of dosShell(), which is platform-dependent, and was not even well implemented, because the __unix__ macro is not defined in all UNIX systems.

The Pascal version of Turbo Vision 2.0 not only defines these functions, but also 'WriteShellMsg' and 'GetTileRect' to control their behaviour, and handleEvent to respond to cmDosShell, cmCascade and cmTile. All of this has been added back to TApplication.

This may break existing applications, because many of them handle these commands manually.