tvforms: do not avoid padding in TDataRec
The reason for this is that TDescrRec mimics TMemoData, which is part of the library API and uses ushort as length type. So, replacing ushort with uint does not solve the issue but rather breaks the expectations of TMemo::getData/setData. Because the padding is located at the end of the struct, getData/setData still work.