TListViewer: fix cursor being shown despite focused item not visible
To reproduce the issue, get to the Open File dialog, focus the penultimate item from the second column and resize the dialog so that it is one row shorter. The focused item is no longer in sight, but because the former cursor position is not out of bounds, the cursor is still visible.

This can be solved by moving the caret out of bounds whenever the focused item is not visible. This will prevent it from being drawn without altering the sfCursorVis state.