platform: fix inconsistencies in cursor size
Implementations of setCursorSize() used to have an immediate effect, both in NcursesDisplay and in Win32Display. This breaks the expectation that all the screen upates are performed at once, and may also cause inconsistencies during drawing (e.g. the cursor becoming visible too early).

This can be solved by caching the caret size in a member variable in BufferedDisplay and then invoke the lower level API during flushScreen().  For this purpose, the lowlevelCursorSize() virtual function has been added to the BufferedDisplay interface.