platform: fix initial cursor size on Windows
6c0745c2947f54f8ed150c34cb90653d02bb224b introduced the assumption in BufferedDisplay that the cursor is initially hidden, but this is not true on Windows.

A possible solution is to do something similar to what we did for the screen size: we store in DisplayStrategy a variable that represents the actual cursor size, and update it in DisplayStrategy::reloadScreenInfo(). BufferedDisplay now stores the 'future' cursor size instead of the 'last' one.

The only difference between this solution and the one for screen size is that while DisplayStrategy::size is modified exclusively from DisplayStrategy::reloadScreenInfo(), DisplayStrategy::caretSize is also modified from BufferedDisplay::flushScreen(), because flushScreen() does change the terminal's cursor size.