platform: implement NcursesDisplay::getCaretSize
Well... in 188009b9e6c58e91bd6861d498bc302c3034b0be I mentioned that the initial cursor size cannot be assumed to be zero on Windows, and neither can on Linux! Tht commit, which solved the caret being visible by default on Windows, regressed that same behaviour on Linux. But solving this is now as easy as just implementing getCaretSize.