platform: ensure result of PlatformStrategy::getCaretSize() is greater than 1
This has to be in the range 1 to 100. The result of this function will gets assigned to TScreen::cursorLines and be then used by TView::resetCursor(). If it has the wrong value, the cursor won't be displayed correctly. Regressed by 188009b9e6c58e91bd6861d498bc302c3034b0be.