BufferedDisplay: use a second buffer to improve performance
It turns out that the performance impact of BufferedDisplay::screenWrite is much larger than BufferedDisplay::flushScreen, because it is invoked way more times. So marking cells as dirty in screenWrite in order to simplify flushScreen results in a pessimization.

screenWrite now does little more than memcpy. Validating cells and copying them to the second buffer is done during flushScreen. This has the advantage that we no longer need to store a dirty bit in TScreenCell, which makes cell comparisons more efficient.

We could optimize this further by not allocating a second buffer and relying on TScreen::screenBuffer directly, because we know TView::writeView copies cells there before invoking screenWrite. But this would make BufferedDisplay harder to understand, so I prefer not going that way unless allocating two buffers becomes a problem.

This commit also fixes a pessimization in TDrawBuffer::moveChar.