platform: reorganize detection of terminal capabilities
Add a new class to the DisplayStrategy hierarchy, TerminalDisplay, representing a strategy which depends on terminal capabilities.

Allow each display strategy (NcursesDisplay, Win32ConsoleDisplay) to provide information about the supported colors and attribute conversion quirks. Store this information in a member variable of TerminalDisplay and initialize it during 'reloadScreenInfo'.

This makes it possible to deal with terminal capabilities in a cross-platform way and without duplicating logic.